Refactored user routes to separate module
